Rename result data component
